creating custom hook for fetching projects using react-query
